move types to helpers and use @frinx path instead of ../../../
